Correct squarenorm() incantations in array.h and matrix.h
Former-commit-id: 1d7b54f7bc5385176ffd6624e3c2a01c8bf3b043

+++ applications/benchmarks/bench_matrix2.cpp
@@ -43,7 +43,7 @@ const CoordinateVector latsize = {32, 32, 32, 32};
 // from msize.
 ///////////////////////////////////////
 
-using ntype = float;
+using ntype = double;
 
 int main(int argc, char **argv) {
     int n_runs = 1;
